+++ src/controllers/epsode.controller.ts
@@ -1,22 +1,28 @@
 import { Request, Response } from "express";
 import { epsodeRepository } from "../repositories/epsode.repository";
 import { serieRepository } from "../repositories/serie.repository";
+import { directorRepository } from "../repositories/directory.repository";
 
 export class epsodeController {
     async create(req: Request, res: Response) {
-        const { title, synopsis, duration, serie } = req.body
+        const { title, synopsis, duration, serie, director  } = req.body
         const serieId = await serieRepository.findOneBy({ id: serie })
+        const directorId = await directorRepository.findOneBy({ id: director })
+
+        const titleDb = await epsodeRepository.findOneBy({title: title})
+        if(titleDb) return res.status(400).json({message: "Title duplicated"})
 
         try {
             const newEpsode = epsodeRepository.create({
                 title,
                 synopsis,
                 duration,
-                serie: { id: serieId?.id }
+                serie: { id: serieId?.id },
+                director: { id: directorId?.id }
             })
             await epsodeRepository.save(newEpsode)
 
-            return res.status(201).json({ message: "Epsode added", data: { ...newEpsode, serie: serieId } })
+            return res.status(201).json({ message: "Epsode added", data: { ...newEpsode, serie: serieId, director: directorId } })
         } catch (error) {
             console.log(error);
             return res.status(500).json({ message: "Internal server error" })
@@ -26,7 +32,7 @@ export class epsodeController {
     async findAll(req: Request, res: Response) {
         try {
             const takeAll = await epsodeRepository.find({
-                relations: ["serie"]
+                relations: ["serie", "director"]
             })
 
             return res.status(200).json({ message: takeAll })
@@ -40,7 +46,7 @@ export class epsodeController {
         const { epsodeId } = req.params
         try {
             const epsode = await epsodeRepository.findOne({
-                where: { id: epsodeId }, relations: ["serie"]
+                where: { id: epsodeId }, relations: ["serie", "director"]
             })
 
             if (!epsode) return res.status(404).json({ message: "Epsode not found" })
@@ -54,9 +60,15 @@ export class epsodeController {
 
     async update(req: Request, res: Response) {
         const { epsodeId } = req.params
-        const { title, synopsis, duration, serie } = req.body
+        const { title, synopsis, duration, serie, director } = req.body
 
         const serieId = await serieRepository.findOneBy({ id: serie })
+        const episodeId = await epsodeRepository.findOneBy({ id: epsodeId })
+        const directorId = await directorRepository.findOneBy({ id: director })
+        
+        if(!serieId) return res.status(400).json({message: "serie not found"})
+        if(!directorId) return res.status(400).json({message: "director not found"})
+        if(!episodeId) return res.status(400).json({message: "episode not found"})
 
         try {
             await epsodeRepository.update({
@@ -83,6 +95,10 @@ export class epsodeController {
 
     async exclude(req: Request, res: Response) {
         const { epsodeId } = req.params
+
+        const episodeId = await epsodeRepository.findOneBy({ id: epsodeId })
+
+        if(!episodeId) return res.status(400).json({message: "episode not found"})
 
         try {
             await epsodeRepository.delete({ id: epsodeId })
